feat(kotori): modern JS parser fixes — Google.com now parses clean
14 fixes for real-world JS compatibility, tested against Google.com's
1MB+ minified scripts. Parse errors went from 8 to 0.

Parser:
- Comma operator support (parseExpression at comma precedence)
- Class expressions in expression position
- super keyword in parsePrefix
- Array elision [,...rest] handling
- Async arrow functions: async () => {} and async x => {}
- Rest params in grouped expressions: (...x) => {}
- Spread in new() constructor args: new X(...args)
- Generator function expressions: function*() {}
- new.target meta-property
- Class method params use parseFunctionParam (rest/defaults)
- 'of' treated as identifier outside for-of context

Lexer:
- ?. vs ? .65 disambiguation (ternary + decimal number)
- Regex character class [/] — don't treat / inside [] as terminator
- Token.len u16→u32 for >64KB string literals

Compiler:
- Comma operator codegen: eval lhs, pop, eval rhs

+++ src/js/kotori/token.zig
@@ -137,7 +137,7 @@ pub const TokenType = enum(u8) {
 pub const Token = struct {
     type: TokenType,
     start: u32,
-    len: u16,
+    len: u32,
     line: u32,
 
     pub fn slice(self: Token, source: []const u8) []const u8 {
